Clean up the grammar rules and add testing
Rather than parsing a single file, the test executable takes an entire
collection of test source code files which aren't hard-linked in the
source but the user can add new test files without changing the source
code of the test executable. Each test source should be coupled with
the expected AST output given in an unambiguous and humanly
readable/writable format.

* Stabilize the grammar just before adding testing to the parser. It
must parse the given test source and produce proper AST for it.
* Turn the StorageClass lexer rule to parser rule because it's causing
issues with the static imports.
* Simplify the module declaration rule without rewrite syntax
(specifying the root node).
* Remove the ambigous parts of imports and the moduleFullyQualifiedName
rules.
* Read the lines from the source and compare the lines to the
CommonTree.toStringTree() results to the previous comment line without
the comment prefix.
* Print the CommonTree node for the that node that has no AST
description
in the previous comment.
* Change the rewrite rule's first child node of the 'type' parser rule
to
primaryExpression, as the current rule adds a false IntegerLiteral with
no available token information. This is causing a warning message in the
test, as the comment line can't be found with the codeLine returning 0.
* Change the package name to dev.progician.dparser and the grammar to
Dee.g which will produce DeeParser and DeeLexer respectively.
* Show '[pass]' for each test suites/files that passed the checking.
* Skip extra spaces in the AST description.